move merge header test out of header Normaliziation Test and update
