+++ repose-aggregator/functional-tests/spock-functional-test/src/test/configs/features/filters/mergeheader/system-model.cfg.xml
@@ -8,7 +8,6 @@
 
         <filters>
             <filter name="merge-header"/>
-            <filter name="header-normalization"/>
         </filters>
 
         <destinations>
